requirements.txt
# ==========================================
# SYPNOSIS ENGINE 1 - Verified Environment
# Hardware: AMD Radeon (ROCm 6.0.0)
# Python: 3.12.3
# ==========================================

# --- AMD GPU Acceleration (ROCm 6.0) ---
# CRITICAL: Do not remove the +rocm6.0 suffix
torch==2.4.1+rocm6.0
torchvision==0.19.1+rocm6.0
torchaudio==2.4.1+rocm6.0
pytorch-triton-rocm==3.0.0

# --- OCR Engines ---
paddleocr==2.7.0.2
paddlepaddle==3.3.0
easyocr==1.7.2
numpy==1.26.4

# --- Vector Database & AI ---
chromadb==1.5.1
ollama==0.6.1
ImageHash==4.3.2

# --- Automation & Screen Monitoring ---
PyAutoGUI==0.9.54
python3-xlib==0.15
psutil==7.2.2

# --- Vision & Processing ---
opencv-python==4.6.0.66
pillow==12.1.1
scipy==1.17.0
scikit-image==0.26.0

# --- Utilities ---
python-dotenv==1.2.1
requests==2.32.5
tqdm==4.67.3
rich==14.3.2